一覧側もリファクタリング
