+++ frontend/features/Colleagues/index.tsx
@@ -1,27 +1,11 @@
 "use client";
 
-import { useEffect, useState } from "react";
 import { Presenter } from "./Presenter";
-import { apiClient } from "@/libs/http/apiClient";
-import { useRouter } from "next/navigation";
 import { Box, CircularProgress } from "@mui/material";
-import { TeachersData } from "./types";
+import { useColleagues } from "./hooks";
 
 export const Colleagues = () => {
-  const [data, setData] = useState<TeachersData | null>(null);
-  const [page, setPage] = useState(1);
-  const router = useRouter();
-
-  useEffect(() => {
-    apiClient
-      .get<TeachersData>("/api/teacher/colleagues")
-      .then((res) => setData(res.data))
-      .catch((err) => {
-        if (err.response?.status === 401) {
-          router.push("/login");
-        }
-      });
-  }, [router]);
+  const { data, page, setPage } = useColleagues();
 
   if (!data) {
     return (
